+++ handler/api/src/main/java/io/knotx/fragments/handler/api/domain/FragmentResult.java
@@ -15,7 +15,7 @@
  */
 package io.knotx.fragments.handler.api.domain;
 
-import io.knotx.fragment.Fragment;
+import io.knotx.fragments.api.Fragment;
 import io.vertx.codegen.annotations.DataObject;
 import io.vertx.core.json.JsonObject;
 import java.util.Objects;
