+++ handler/core/src/main/java/io/knotx/fragments/handler/FragmentsHandler.java
@@ -19,14 +19,7 @@ package io.knotx.fragments.handler;
 
 import static com.google.common.base.Predicates.alwaysTrue;
 
-import java.util.Iterator;
-import java.util.List;
-import java.util.ServiceLoader;
-import java.util.function.Predicate;
-import java.util.function.Supplier;
-import java.util.stream.Collectors;
-
-import io.knotx.fragment.Fragment;
+import io.knotx.fragments.api.Fragment;
 import io.knotx.fragments.engine.FragmentEvent;
 import io.knotx.fragments.engine.FragmentEvent.Status;
 import io.knotx.fragments.engine.FragmentEventContext;
@@ -47,6 +40,12 @@ import io.vertx.core.Handler;
 import io.vertx.core.json.JsonObject;
 import io.vertx.reactivex.core.Vertx;
 import io.vertx.reactivex.ext.web.RoutingContext;
+import java.util.Iterator;
+import java.util.List;
+import java.util.ServiceLoader;
+import java.util.function.Predicate;
+import java.util.function.Supplier;
+import java.util.stream.Collectors;
 
 public class FragmentsHandler implements Handler<RoutingContext> {
