+++ handler/core/src/main/java/io/knotx/fragments/handler/action/InlinePayloadActionFactory.java
@@ -15,7 +15,7 @@
  */
 package io.knotx.fragments.handler.action;
 
-import io.knotx.fragment.Fragment;
+import io.knotx.fragments.api.Fragment;
 import io.knotx.fragments.handler.api.Action;
 import io.knotx.fragments.handler.api.ActionFactory;
 import io.knotx.fragments.handler.api.Cacheable;
