+++ handler/engine/README.md
@@ -1,6 +1,6 @@
 # Fragments Engine
 This module is a heart of the *Fragment Processing*. It actually does the work of independent 
-[**Fragments**](https://github.com/Knotx/knotx-fragment-api#knotx-fragment-api) processing,
+[**Fragments**](https://github.com/Knotx/knotx-fragments/tree/master/api) processing,
 by routing each *Fragment* throught the graph and applying [**Actions**](https://github.com/Knotx/knotx-fragments-handler/tree/master/api#action)
  during that processing.
 
@@ -58,23 +58,23 @@ The images below illustrates the above rules.
 
 * *Node A* and *Node B* ends correctly:
 
-![Node with exits](assets/images/a_next_b.png)
+![Node with exits](../assets/images/a_next_b.png)
 
 * *Node A* raises an exception, *Node B* ends correctly 
 
-![Node with exits](assets/images/a_error_c.png)
+![Node with exits](../assets/images/a_error_c.png)
 
 `FAILURE` states:
 
 * *Node A* ends correctly and *Node B* raises exception:
 
-![Node with exits](assets/images/a_next_b_error.png)
+![Node with exits](../assets/images/a_next_b_error.png)
 
 * *Node A* and *Node C* raises exceptions:
 
-![Node with exits](assets/images/a_error_c_error.png)
+![Node with exits](../assets/images/a_error_c_error.png)
 
 A node can also declare its own exits (transitions) but then we need to configure them in a graph. 
 Otherwise, if the custom transition is set but is not declared, then the `FAILURE` status is returned:
 
-![Node with exits](assets/images/a_custom.png)
+![Node with exits](../assets/images/a_custom.png)
