+++ handler/engine/src/test/java/io/knotx/fragments/engine/FragmentsEngineOrderTest.java
@@ -18,9 +18,9 @@ package io.knotx.fragments.engine;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import io.knotx.fragment.Fragment;
-import io.knotx.fragments.engine.graph.Node;
+import io.knotx.fragments.api.Fragment;
 import io.knotx.fragments.engine.graph.ActionNode;
+import io.knotx.fragments.engine.graph.Node;
 import io.knotx.fragments.handler.api.domain.FragmentContext;
 import io.knotx.fragments.handler.api.domain.FragmentResult;
 import io.knotx.server.api.context.ClientRequest;
