+++ handler/engine/src/test/java/io/knotx/fragments/engine/TaskEngineSingleOperationTest.java
@@ -25,7 +25,7 @@ import static io.knotx.fragments.handler.api.domain.FragmentResult.ERROR_TRANSIT
 import static io.knotx.fragments.handler.api.domain.FragmentResult.SUCCESS_TRANSITION;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import io.knotx.fragment.Fragment;
+import io.knotx.fragments.api.Fragment;
 import io.knotx.fragments.engine.FragmentEvent.Status;
 import io.knotx.fragments.engine.FragmentEventLogVerifier.Operation;
 import io.knotx.fragments.engine.graph.ActionNode;
