+++ supplier/api/src/main/java/io/knotx/fragments/supplier/api/FragmentsSupplier.java
@@ -15,7 +15,7 @@
  */
 package io.knotx.fragments.supplier.api;
 
-import io.knotx.fragment.Fragment;
+import io.knotx.fragments.api.Fragment;
 import io.knotx.server.api.context.RequestContext;
 import java.util.List;
